// backend/src/common/enums/resource-category.enum.ts
export enum ResourceCategory {
  DESKTOP_COMPUTER = 'DESKTOP_COMPUTER',
  LAPTOP = 'LAPTOP',
  SERVER = 'SERVER',
  MOBILE_DEVICE = 'MOBILE_DEVICE',
  TABLET = 'TABLET',
  PERIPHERAL = 'PERIPHERAL',
  NETWORK_EQUIPMENT = 'NETWORK_EQUIPMENT',
  SOFTWARE_LICENSE = 'SOFTWARE_LICENSE',
  CLOUD_SERVICE = 'CLOUD_SERVICE',
  OTHER = 'OTHER',
}